requirements.txt
crewai>=0.11.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0 